cheri: Apply the freepage.cocci semantic patch
Calls to free_page and free_pages that cast its (pointer)
argument to unsigned long must cast it to uintptr_t instead.

Apply like this:

	make COCCI=scripts/coccinelle/cheri/free_page.cocci \
		MODE=patch coccicheck

+++ arch/s390/kernel/perf_pai_crypto.c
@@ -83,7 +83,7 @@ static DEFINE_MUTEX(pai_reserve_mutex);
 /* Free all memory allocated for event counting/sampling setup */
 static void paicrypt_free(struct paicrypt_mapptr *mp)
 {
-	free_page((unsigned long)mp->mapptr->page);
+	free_page((uintptr_t)mp->mapptr->page);
 	kvfree(mp->mapptr->save);
 	kfree(mp->mapptr);
 	mp->mapptr = NULL;
