+++ arch/s390/kernel/smp.c
@@ -217,7 +217,7 @@ out:
 	stack_free(mcck_stack);
 	stack_free(async_stack);
 	free_pages(nodat_stack, THREAD_SIZE_ORDER);
-	free_pages((unsigned long) lc, LC_ORDER);
+	free_pages((uintptr_t) lc, LC_ORDER);
 	return -ENOMEM;
 }
 
@@ -237,7 +237,7 @@ static void pcpu_free_lowcore(struct pcpu *pcpu, int cpu)
 	stack_free(async_stack);
 	stack_free(mcck_stack);
 	free_pages(nodat_stack, THREAD_SIZE_ORDER);
-	free_pages((unsigned long) lc, LC_ORDER);
+	free_pages((uintptr_t) lc, LC_ORDER);
 }
 
 static void pcpu_prepare_secondary(struct pcpu *pcpu, int cpu)
