+++ kernel/power/swap.c
@@ -400,7 +400,7 @@ sync_io:
 static void release_swap_writer(struct swap_map_handle *handle)
 {
 	if (handle->cur)
-		free_page((unsigned long)handle->cur);
+		free_page((uintptr_t)handle->cur);
 	handle->cur = NULL;
 }
 
@@ -903,7 +903,7 @@ out_clean:
 		vfree(data);
 	}
 	if (page)
-		free_page((unsigned long)page);
+		free_page((uintptr_t)page);
 
 	return ret;
 }
@@ -988,7 +988,7 @@ static void release_swap_reader(struct swap_map_handle *handle)
 
 	while (handle->maps) {
 		if (handle->maps->map)
-			free_page((unsigned long)handle->maps->map);
+			free_page((uintptr_t)handle->maps->map);
 		tmp = handle->maps;
 		handle->maps = handle->maps->next;
 		kfree(tmp);
@@ -1062,7 +1062,7 @@ static int swap_read_page(struct swap_map_handle *handle, void *buf,
 		return error;
 	if (++handle->k >= MAP_PAGE_ENTRIES) {
 		handle->k = 0;
-		free_page((unsigned long)handle->maps->map);
+		free_page((uintptr_t)handle->maps->map);
 		tmp = handle->maps;
 		handle->maps = handle->maps->next;
 		kfree(tmp);
@@ -1509,7 +1509,7 @@ out_finish:
 out_clean:
 	hib_finish_batch(&hb);
 	for (i = 0; i < ring_size; i++)
-		free_page((unsigned long)page[i]);
+		free_page((uintptr_t)page[i]);
 	if (crc) {
 		if (crc->thr)
 			kthread_stop(crc->thr);
